fix: fixes #700
