+++ components/inferno/index.d.ts
@@ -4,28 +4,19 @@
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
 /// <reference types="inferno" />
+import type { ParticlesProps } from "./src/";
 import type { IOptions } from "tsparticles/dist/Options/Interfaces/IOptions";
 import type { RecursivePartial } from "tsparticles/dist/Types/RecursivePartial";
-import { IPolygonMaskOptions } from "tsparticles/dist/Plugins/PolygonMask/PolygonMaskPlugin";
-import { IAbsorberOptions } from "tsparticles/dist/Plugins/Absorbers/AbsorbersPlugin";
-import { IEmitterOptions } from "tsparticles/dist/Plugins/Emitters/EmittersPlugin";
-import { Component, RefObject } from "inferno";
-import { Container } from "tsparticles/dist/Core/Container";
+import type { IPolygonMaskOptions } from "tsparticles/dist/Plugins/PolygonMask/PolygonMaskPlugin";
+import type { IAbsorberOptions } from "tsparticles/dist/Plugins/Absorbers/AbsorbersPlugin";
+import type { IEmitterOptions } from "tsparticles/dist/Plugins/Emitters/EmittersPlugin";
+import type { Component } from "inferno";
 
 export type IParticlesParams = RecursivePartial<IOptions & IPolygonMaskOptions & IAbsorberOptions & IEmitterOptions>;
 
 export * from "tsparticles/dist/Enums";
 
-export interface ParticlesProps {
-	width?: string;
-	height?: string;
-	params?: IParticlesParams;
-	options?: IParticlesParams;
-	style?: any;
-	className?: string;
-	canvasClassName?: string;
-	container?: RefObject<Container>;
-}
+export { ParticlesProps };
 
 type Particles = Component<ParticlesProps>;
