+++ components/inferno/src/index.ts
@@ -2,5 +2,8 @@ import Particles from "./Particles";
 
 export * from "tsparticles/dist/Enums";
 
+export { ParticlesProps } from "./ParticlesProps";
+export { ParticlesState } from "./ParticlesState";
+
 export default Particles;
 export { Particles };
