+++ components/react/index.d.ts
@@ -5,27 +5,18 @@
 
 /// <reference types="react" />
 import { ComponentClass } from "react";
-import { Container } from "tsparticles/dist/Core/Container";
+import { ParticlesProps } from "./src/";
 import type { IOptions } from "tsparticles/dist/Options/Interfaces/IOptions";
 import type { RecursivePartial } from "tsparticles/dist/Types/RecursivePartial";
-import { IPolygonMaskOptions } from "tsparticles/dist/Plugins/PolygonMask/PolygonMaskPlugin";
-import { IAbsorberOptions } from "tsparticles/dist/Plugins/Absorbers/AbsorbersPlugin";
-import { IEmitterOptions } from "tsparticles/dist/Plugins/Emitters/EmittersPlugin";
+import type { IPolygonMaskOptions } from "tsparticles/dist/Plugins/PolygonMask/PolygonMaskPlugin";
+import type { IAbsorberOptions } from "tsparticles/dist/Plugins/Absorbers/AbsorbersPlugin";
+import type { IEmitterOptions } from "tsparticles/dist/Plugins/Emitters/EmittersPlugin";
 
 export type IParticlesParams = RecursivePartial<IOptions & IPolygonMaskOptions & IAbsorberOptions & IEmitterOptions>;
 
 export * from "tsparticles/dist/Enums";
 
-export interface ParticlesProps {
-    width?: string;
-    height?: string;
-    params?: IParticlesParams;
-    options?: IParticlesParams;
-    style?: any;
-    className?: string;
-    canvasClassName?: string;
-    container?: React.RefObject<Container>;
-}
+export { ParticlesProps };
 
 type Particles = ComponentClass<ParticlesProps>;
