+++ components/react/src/ParticlesState.ts
@@ -1,0 +1,6 @@
+import type { Container } from "tsparticles/dist/Core/Container";
+
+export interface ParticlesState {
+    canvas?: HTMLCanvasElement;
+    library?: Container;
+}
