+++ core/main/src/Core/Particle/Interactions/Mouse/Grabber.ts
@@ -8,8 +8,7 @@ import { IExternalInteractor } from "../../../Interfaces/IExternalInteractor";
  * Particle grab manager
  */
 export class Grabber implements IExternalInteractor {
-    constructor(private readonly container: Container) {
-    }
+    constructor(private readonly container: Container) {}
 
     public isEnabled(): boolean {
         const container = this.container;
